Added account route to the infomation blueprint
