+++ api/templates/_topnavigation.html
@@ -2,6 +2,6 @@
 <a href="{{ url_for('info.index') }}" class="logo">QUIZIVERSE</a>
 <nav>
   <a href="{{ url_for('dash.home') }}">Home</a>
-  <a href="{{ url_for('account') }}">Account</a>
+  <a href="{{ url_for('info.account') }}">Account</a>
   <a href="{{ url_for('logout') }}">Logout</a>
 </nav>
